+++ galsim/random.py
@@ -492,11 +492,97 @@ _galsim.Chi2Deviate.setN.__func__.__doc__ = "Set current distribution n degrees 
 
 #
 # The rest of these are from Noise.h in the C++ layer, not Randome.h.
+#
 
 
-# CCDNoise deviate docstrings
+# GaussianNoise docstrings
+_galsim.GaussianNoise.__doc__ = """
+Class implementing simple Gaussian noise.
+
+This is a simple noise model where each pixel in the image gets Gaussian noise with a 
+given sigma.
+
+Initialization
+--------------
+
+    >>> gaussian_noise = galsim.GaussianNoise(rng, sigma=1.)
+
+Parameters:
+
+    rng       A BaseDeviate instance to use for generating the random numbers.
+    sigma     The rms noise on each pixel [default `sigma = 1.`].
+
+Methods
+-------
+To add noise to every element of an image, use the syntax image.addNoise(gaussian_noise).
+"""
+
+_galsim.GaussianNoise.applyTo.__func__.__doc__ = """
+Add Gaussian noise to an input Image.
+
+Calling
+-------
+
+    >>> gaussian_noise.applyTo(image)
+
+On output the Image instance image will have been given additional Gaussian noise according 
+to the given GaussianNoise instance.
+
+Note: The syntax image.addNoise(gaussian_noise) is preferred.
+"""
+_galsim.GaussianNoise.getSigma.__func__.__doc__ = "Get sigma in current noise model."
+_galsim.GaussianNoise.setSigma.__func__.__doc__ = "Set sigma in current noise model."
+
+
+# PoissonNoise docstrings
+_galsim.PoissonNoise.__doc__ = """
+Class implementing Poisson noise according to the flux (and sky level) present in the image.
+
+The PoissonNoise class encapsulates a simple version of the noise model of a normal CCD image
+where each pixel has Poisson noise corresponding to the number of electrons in each pixel
+(including an optional extra sky level).
+
+Note that if the image to which you are adding noise already has a sky level on it,
+then you should not provide the sky level here as well.  The sky level here corresponds
+to a level is taken to be already subtracted from the image, but which was present
+for the Poisson noise.
+
+Initialization
+--------------
+
+    >>> poisson_noise = galsim.PoissonNoise(rng, sky_level=0.)
+
+Parameters:
+
+    rng         A BaseDeviate instance to use for generating the random numbers.
+    sky_level   The sky level in electrons per pixel that was originally in the input image, 
+                but which is taken to have already been subtracted off [default `sky_level = 0.`].
+
+Methods
+-------
+To add noise to every element of an image, use the syntax image.addNoise(poisson_noise).
+"""
+
+_galsim.PoissonNoise.applyTo.__func__.__doc__ = """
+Add Poisson noise to an input Image.
+
+Calling
+-------
+
+    >>> galsim.PoissonNoise.applyTo(image)
+
+On output the Image instance image will have been given additional Poisson noise according 
+to the given PoissonNoise instance.
+
+Note: the syntax image.addNoise(poisson_noise) is preferred.
+"""
+_galsim.PoissonNoise.getSkyLevel.__func__.__doc__ = "Get sky level in current noise model."
+_galsim.PoissonNoise.setSkyLevel.__func__.__doc__ = "Set sky level in current noise model."
+
+
+# CCDNoise docstrings
 _galsim.CCDNoise.__doc__ = """
-Pseudo-random number generator with a basic CCD noise model.
+Class implementing a basic CCD noise model.
 
 The CCDNoise class encapsulates the noise model of a normal CCD image.  The noise has two
 components: first, Poisson noise corresponding to the number of electrons in each pixel
@@ -525,10 +611,7 @@ Parameters:
 
 Methods
 -------
-To add deviates to every element of an image, use the syntax image.addNoise(ccd_noise).
-
-To get and set the deviate parameters, see the docstrings for the getSkyLevel(), getGain(), 
-getReadNoise(), setSkyLevel(), setGain(), and setReadNoise() methods of each instance.
+To add noise to every element of an image, use the syntax image.addNoise(ccd_noise).
 """
 
 _galsim.CCDNoise.applyTo.__func__.__doc__ = """
@@ -537,12 +620,12 @@ Add noise to an input Image.
 Calling
 -------
 
-    >>> galsim.CCDNoise.applyTo(image)
+    >>> ccd_noise.applyTo(image)
 
-On output the Image instance image will have been given an additional stochastic noise according to 
-the gain and read noise settings of the CCDNoise instance.
+On output the Image instance image will have been given additional stochastic noise according to 
+the gain and read noise settings of the given CCDNoise instance.
 
-To add deviates to every element of an image, the syntax image.addNoise() is preferred.
+Note: the syntax image.addNoise(ccd_noise) is preferred.
 """
 _galsim.CCDNoise.getSkyLevel.__func__.__doc__ = "Get sky level in current noise model."
 _galsim.CCDNoise.getGain.__func__.__doc__ = "Get gain in current noise model."
@@ -551,4 +634,39 @@ _galsim.CCDNoise.setSkyLevel.__func__.__doc__ = "Set sky level in current noise 
 _galsim.CCDNoise.setGain.__func__.__doc__ = "Set gain in current noise model."
 _galsim.CCDNoise.setReadNoise.__func__.__doc__ = "Set read noise in current noise model."
 
+
+# DeviateNoise docstrings
+_galsim.DeviateNoise.__doc__ = """
+Class implementing noise with an arbitrary BaseDeviate object.
+
+The DeviateNoise class provides a way to treat an arbitrary deviate as the noise model for 
+each pixel in an image.
+
+Initialization
+--------------
+
+    >>> dev_noise = galsim.DeviateNoise(dev)
+
+Parameters:
+
+    dev         A BaseDeviate subclass to use as the noise deviate for each pixel.
+
+Methods
+-------
+To add noise to every element of an image, use the syntax image.addNoise(dev_noise).
+"""
+
+_galsim.DeviateNoise.applyTo.__func__.__doc__ = """
+Add noise to an input Image.
+
+Calling
+-------
+
+    >>> dev_noise.applyTo(image)
+
+On output the Image instance image will have been given additional noise according to 
+the given DeviateNoise instance.
+
+To add deviates to every element of an image, the syntax image.addNoise() is preferred.
+"""