+++ CHANGELOG.md
@@ -17,3 +17,6 @@ Changes from v0.3 to current version:
 
 * Added an addNoiseSNR() method to Image classes that adds noise such that the image has a 
   specified signal-to-noise ratio. (Issue #349)
+
+* Made a new Noise hierarchy, and moved CCDNoise to that rather than have it be a BaseDeviate.
+  There are also now GaussianNoise, PoissonNoise, and DeviateNoise classes. (Issue #349)